refactor: change headers and improve readabilty

+++ ft_striteri.c
@@ -6,13 +6,15 @@
 /*   By: rde-mour <rde-mour@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/12 09:57:08 by rde-mour          #+#    #+#             */
-/*   Updated: 2023/10/29 20:38:11 by rde-mour         ###   ########.org.br   */
+/*   Updated: 2024/09/27 12:10:40 by rde-mour         ###   ########.org.br   */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include <stddef.h>
+
 void	ft_striteri(char *s, void (*f)(unsigned int, char *))
 {
-	unsigned int	i;
+	size_t	i;
 
 	if (!s)
 		return ;
